python-portfolio/requirements.txt
flask
pymongo
cloudinary
python-dotenv